Rename and simplify enumeration method
